% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.chain.R
\name{add.chain}
\alias{add.chain}
\title{Adding chain to custom traceplot}
\usage{
add.chain(trace.vec, thin = 0, col = "black")
}
\arguments{
\item{trace.vec}{Vector containing draws from the MCMC sampler}

\item{thin}{Amount of thinning applied to the plot.  For example, if thin=3,
then every third sample point will be plotted.  The total number of
iterations should be divisible by the thinning amount.}

\item{col}{Color of trace line. Defaults to black.}
}
\description{
Function to add a trace line for a single chain of an MCMC to an existing
custom traceplot created by my.traceplot.
}
\examples{
## Draw a traceplot and add chains for a random sample
my.samples = rnorm(10000, 5, 10)
my.samples2 = rnorm(10000, 5, 20)
my.traceplot(my.samples,label="Some title",thin=5,burn=1000,col="blue")
add.chain(my.samples2,thin=5,col="green")
}
\seealso{
my.traceplot
}
\keyword{chain}
\keyword{traceplot,}

