% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{BCLmat}
\alias{BCLmat}
\title{Convert vector version of baseline category logit in familiar matrix form}
\usage{
BCLmat(coef_vec_M, nm = length(unique(M)))
}
\arguments{
\item{coef_vec_M}{coefficient vector version}

\item{nm}{Number of unique categories for (including reference)}
}
\description{
First row is intercept for each of the categories (minus reference)
First column corresponds to covariate for first non-reference level
}

