% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bunny_function.R
\name{bunny_function}
\alias{bunny_function}
\title{A bunny litmus tests}
\usage{
bunny_function(love = TRUE)
}
\arguments{
\item{love}{Do you love bunnies? Defaults to TRUE.  How could it not?}
}
\description{
By popular demand, this function allows you to express your love of bunnies.
This is a variant of Hilary Parker's example function, cat_function.
}
\examples{
bunny_function()
}
\seealso{
cat_function
}
\keyword{bunnies,}
\keyword{bunny,}
\keyword{rabbit,}
\keyword{rabbits}

