% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
This is Hilary Parker's example function.
For some reason, my projects don't really use it too often.
Perhaps I am studying the wrong kind of statistics.
}
\examples{
cat_function()
}
\keyword{cats}

