% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/col2alpha.R
\name{col2alpha}
\alias{col2alpha}
\title{Quickly add alpha transparency to a color}
\usage{
col2alpha(col, alpha = 0.5)
}
\arguments{
\item{col}{Opaque color. Can be named ("red") or hexadecimal, as in what
RColorBrewer returns.}

\item{alpha}{Alpha transparency, where 1 is opaque and 0 is transparent.
Defaults to 0.5.}
}
\description{
Function to return color value with transparency added
}
\examples{
## Demonstration with named color
plot(1:3,1:3, col=col2alpha("red",0.8), type="b", lwd=2)
lines(3:1,1:3, col=col2alpha("red",0.2), lwd=4)
}
\seealso{
col2rgb
}
\keyword{alpha,}
\keyword{rgb}
\keyword{transparency,}

