% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/colMedians.R
\name{colMedians}
\alias{colMedians}
\title{colMedians}
\usage{
colMedians(colMedians.object, na.rm = FALSE)
}
\arguments{
\item{colMedians.object}{Object for which we wish to calculate column medians}

\item{na.rm}{Should NAs be excluded? Defaults to FALSE}
}
\description{
Function to calculate the column-wise medians of a matrix of data frame
(like a somewhat-simplified median analog to colMeans).
Useful for calculating posterior medians in Bayesian statistics.
}
\examples{
## Compute medians for a matrix
my.matrix = matrix(1:15,3,5)
colMedians(my.matrix)
## Computer medians for a data frame
my.df = as.data.frame(my.matrix)
colMedians(my.df)
}
\keyword{median}

