% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R, R/RcppExports.R
\name{column_picker}
\alias{column_picker}
\title{Choose column based on an input vector}
\usage{
column_picker(M, v)

column_picker(M, v)
}
\arguments{
\item{M}{The input matrix}

\item{v}{The length-n vector taking integer values 1 to m}

\item{column}{The vector containing the columns to be extracted}

\item{M}{The n by m matrix}
}
\description{
Choose column based on an input vector

From a matrix and a vector, return a vector where the ith element is the
matrix element (i, v[i])
}
\details{
Faster version of M[cbind(1:length(v), v)]
}

