% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{dMultinom}
\alias{dMultinom}
\title{Multinomial density (temporary)}
\usage{
dMultinom(x, probs, log = FALSE)
}
\arguments{
\item{x}{Length-n outcome vector (1:number of possible categories)}

\item{probs}{Matrix of probabilities (nrows = n, ncol = # categories)}

\item{log}{Return log-density?  Defaults to FALSE}
}
\description{
Will change when Rcpp stops being whiney
}

