% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dMultinomCpp}
\alias{dMultinomCpp}
\title{Evaluate vector of multinomial densities}
\usage{
dMultinomCpp(x, prob, lg)
}
\arguments{
\item{x}{Outcome vector, length n}

\item{prob}{Probability matrix.  Entry ij is P(X_i = j)}
}
\description{
Evaluate vector of multinomial densities
}

