% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit (Inverse logit)}
\usage{
expit(x)
}
\arguments{
\item{input}{Argument to the inverse logit function}
}
\description{
Function to calculate the inverse-logit (or expit)
$\textrm{expit}(a) = \frac{1}{1+e^{-a}} = \frac{e^a}{1+e^a}$
}
\examples{
## Single value
expit(5)
## Matrix of values
expit(matrix(1:4,2,2))
}
\keyword{inverse-logit}

