% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanSDCI.R
\name{getMeanSDCI}
\alias{getMeanSDCI}
\title{Make table of posterior mean, standard deviation, and credible interval}
\usage{
getMeanSDCI(getMeanSDCI.df, alpha = 0.05)
}
\arguments{
\item{getMeanSDCI.df}{Data frame whose columns contain MCMC draws}

\item{alpha}{Type I error rate for credible interval.  Defaults to 0.05.}
}
\description{
Function to calculate posterior mean, standard deviation, and (1-alpha)x100% 
credible interval resultle based on a data frame with columns as parameters' 
MCMC draws.  Column names will be taken as parameter labels.
}
\examples{
## Calculate mean, median, and 95\% quantile-based interval for columns
my.df = as.data.frame(matrix(rnorm(5000,rep(1:5,each=1000),0.5),1000,5))
getMeanSDCI(my.df)

}
\seealso{
makeCIs, getPostMeanMedCI
}
\keyword{posterior}
\keyword{summary}

