% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{impute_U}
\alias{impute_U}
\title{Impute unmeasured binary confounder U}
\usage{
impute_U(coef_U, coef_M, coef_Y, Z, A, M, Y)
}
\arguments{
\item{coef_U}{Current values of U regression coefficients}

\item{coef_M}{Current values of M regression coefficients}

\item{coef_Y}{Current values of Y regression coefficients}

\item{Z}{Confounder (matrix)}

\item{A}{Treatment/exposure}

\item{M}{Mediator}

\item{Y}{Outcome}
}
\description{
Uses M, U, and Y parts of the likelihood to impute appropriate U
}

