% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{initialize_containers}
\alias{initialize_containers}
\title{Function to initialize containers}
\usage{
initialize_containers(R, coef_U, coef_M, coef_Y)
}
\arguments{
\item{R}{Number of MCMC scans}

\item{coef_U}{Vector containing seed values for U coefficients}

\item{coef_M}{Matrix containing seed values for M coefficients}

\item{coef_Y}{Vector containing seed values for Y coefficients}
}
\description{
Initialize containters for regression coefficients and acceptance
}

