% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.if.R
\name{inst.if}
\alias{inst.if}
\title{Install and load packages if necessary}
\usage{
inst.if(pkgname)
}
\arguments{
\item{pkgname}{The name of the desired package as a string}
}
\description{
This function checks to see if a package is installed yet
If not, it installs it.  Either way, once installed, it loads the library.
}
\examples{
inst.if("Matrix")

}

