% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{make_dummy}
\alias{make_dummy}
\title{Make a dummy matrix version of a vector (reference level = 1)}
\usage{
make_dummy(vec)
}
\arguments{
\item{vec}{Vector containing values from 1 to k, where k is # of levels}
}
\description{
Make a dummy matrix version of a vector (reference level = 1)
}

