% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBlocks.R
\name{makeBlocks}
\alias{makeBlocks}
\title{Function to make block identifiers for parameter updating}
\usage{
makeBlocks(toBlock, method)
}
\arguments{
\item{toBlock}{Vector, matrix, or array of parameters to be updated
in blocks}

\item{method}{Method for breaking the updating into blocks. 
Options are:
  "individual" (every parameter updated individually), 
  "together" (entire block is updated at once), 
  "rows" (each row of a matrix is updated as a unit),
  "cols" (each row of a matrix is updated as a unit)}
}
\value{
Object of same size and shape as toBlock with entries
corresponding to which update block the parameter is in
}
\description{
Row and column options are for matrices only at the moment; use
apply to block arrays by row or column.
}

