% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeCI.R
\name{makeCI}
\alias{makeCI}
\title{Quantile-based credible interval for a vector}
\usage{
makeCI(makeCI.vec, alpha = 0.05, na.rm = FALSE)
}
\arguments{
\item{makeCI.vec}{Vector for which quantiles will be found}

\item{alpha}{What Type I error level alpha should be used? Defaults to 0.05}

\item{na.rm}{Should NAs be excluded? Defaults to FALSE}
}
\description{
Function to calculate (two-sided) credible interval based on quantiles of a vector
}
\examples{
##
my.vector = rnorm(n = 1000,mean = 0, sd = 10)
makeCI(my.vector)
}
\seealso{
makeCIs
}
\keyword{credible}
\keyword{interval}

