% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeCIs.R
\name{makeCIs}
\alias{makeCIs}
\title{Quantile-based credible intervals for a columns of a matrix}
\usage{
makeCIs(makeCI.obj, alpha = 0.05, na.rm = FALSE)
}
\arguments{
\item{alpha}{What Type I error level alpha should be used? Defaults to 0.05}

\item{na.rm}{Should NAs be excluded? Defaults to FALSE}

\item{makeCI.mat}{Matrix or data frame for which quantiles will be found}
}
\description{
Function to calculate (two-sided) credible intervals based on quantiles of a
matrix columns.
}
\examples{
## Calculation for a matrix
theta1 = rnorm(n = 1000, mean = 0, sd = 10)
theta2 = rnorm(n = 1000, mean = 0, sd = 10)
my.matrix = cbind(theta1, theta2)
makeCIs(my.matrix)
## Calculation for a data frame
my.df = as.data.frame(my.matrix)
makeCIs(my.df)
}
\seealso{
makeCI
}
\keyword{credible}
\keyword{interval}

