% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mclapply.hack.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{mclapply.hack: forking for Windows}
\usage{
mclapply.hack(..., verbose = FALSE, mc.cores = NULL)
}
\arguments{
\item{verbose}{Should users be warned this is hack-y? Defaults to FALSE.}
}
\description{
This function mimics forking (done with mclapply in Mac or Linux) for the
Windows environment.  Designed to be used just like mclapply.  Credit goes to
Nathan VanHoudnos.
(Extremely) lightly adapted from:
https://github.com/nathanvan/mcmc-in-irt/blob/master/post-10-mclapply-hack.R
## post-10-mclapply.hack.R
##
## Nathan VanHoudnos
## nathanvan AT northwestern FULL STOP edu
## July 14, 2014
## Last Edit:  August 26, 2014
##
## A script to implement a hackish version of
## parallel:mclapply() on Windows machines.
}
\examples{
mclapply.hack()
}
\seealso{
mclapply
}

