% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.acfplot.R
\name{my.acfplot}
\alias{my.acfplot}
\title{Custom autocorrelation plots}
\usage{
my.acfplot(acfplot.vec, label = NULL, thin = 0, burn = 0, lag.max = 100)
}
\arguments{
\item{acfplot.vec}{Vector containing draws from the MCMC sampler}

\item{label}{Parameter label and title for the plot. Default is no label.}

\item{thin}{Amount of thinning applied to the plot.  For example, if thin=3, 
then every third sample point will be plotted.  The total number of 
iterations should be divisible by the thinning amount.}

\item{burn}{Number of iterations for burn-in.}

\item{lag.max}{Maximum lag for which to calculate autocorrelation. Defaults
to 100.}
}
\description{
Function to draw autocorrelation plots the way I like them -- with huge 
lags because my chain mixing sucks.
}
\examples{
## Draw an autocorrelation plot for a random sample
my.samples = rnorm(10000, 5, 10)
my.acfplot(my.samples,label="Some title",thin=5,burn=1000,lag.max=30)

}
\keyword{acf,}
\keyword{autocorrelation}

