% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/my.kable.R
\name{my.kable}
\alias{my.kable}
\title{Customized kable (table-making function for knitr)}
\usage{
my.kable(my.kable.obj, caption = NULL, digits = 4, latexTranslate = FALSE,
  greek = TRUE, prettyNum = TRUE)
}
\arguments{
\item{my.kable.obj}{Object (matrix or data frame) passed in to be kabled}

\item{caption}{Caption for the table. Defaults to no caption.}

\item{digits}{Number of significant digits for numbers in the table. This
sometimes works. Defaults to 4.}

\item{latexTranslate}{Whether or not the latexTranslate function from the
Hmisc package should be used to escape out LaTeX-unfriendly characters.
This tends not to work well for complicated cases. Defaults to FALSE.}

\item{greek}{Argument to latexTranslate function. Again, this only works
sometimes. Defaults to TRUE.}

\item{prettyNum}{Whether prettyNum should be used to align decimals, etc.
Occasionally problematic. Defaults to TRUE.}
}
\description{
Function to adapt kable function from knitr package to meet most of my
table-making needs. Tries to do a lot of things, some of which it does
successfully. Right-aligns numeric columns and left-aligns character
columns, pretties up numbers using prettyNum (if desired), escapes out
LaTeX characters (if desired). Writes the table in LaTeX code assuming
that the booktabs package is being used. Definitely a work in progress.
}
\examples{
## Fake table, an example where greek=TRUE actually works out
my.table = data.frame(ABCs = LETTERS[1:3], beta=1:3, values=10^(-1:-3))
my.kable(my.table, latexTranslate=TRUE, greek=TRUE)
}
\seealso{
kable, latexTranslate
}
\keyword{knitr}
\keyword{tables}

