% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.pbPost.R
\name{my.pbPost}
\alias{my.pbPost}
\title{Quick R pushbullet message}
\usage{
my.pbPost(message, type = "note", title = "R message")
}
\arguments{
\item{message}{Message you would like sent to your pushbullet account.}

\item{type}{Type of message to send. I always use "note," but others may be
possible.}

\item{title}{Title of message. (Note: if you are receiving pushbullets on 
your iPhone, the title will be visible from your lock screen.)}
}
\description{
Function to post to Pushbullet only if certain conditions are met.  This 
allows you to have others run your code on their computers without getting 
messages about not having RPushbullet set up.
For you to receive a message from this, you must set up your Pushbullet 
account with your ".rpushbullet.json" file saved to your home directory.
}
\examples{
## This sends a PB if set up on the computer
my.pbPost("body of test message")

}
\seealso{
pbPost
}
\keyword{error}
\keyword{messages}
\keyword{pushbullet,}

