% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.png.R
\name{my.png}
\alias{my.png}
\title{Quick way to increase resolution for pngs}
\usage{
my.png(filename, width = 6, height = 6, unit = "in", res = 300)
}
\arguments{
\item{filename}{Name of file to be created.  Should include the ".png"
extension}

\item{width}{Width of image file, in units specified by "unit." Defaults to 
6 inches.}

\item{height}{Height of image file, in units specified by "unit." Defaults 
to is 6 inches.}

\item{unit}{Unit of image dimensions. Defaults to inches.}

\item{res}{Resolution of image. Defaults to 300 dpi.}
}
\description{
Function to increase resolution of pngs being created. Helpful for inserting
higher-resolution graphs into a LaTeX document
}
\examples{
## This will save a high-resolution graph in your current working directory.
my.png("ExampleGraph.png")
plot(1:5,1:5, type="b", main="Example graph")
dev.off()

}
\seealso{
png
}
\keyword{png}

