% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.pval.R
\name{my.pval}
\alias{my.pval}
\title{P-value formatting}
\usage{
my.pval(pvals)
}
\arguments{
\item{pvals}{Vector of p-values to format}
}
\description{
Function to format p-values the way I like
}
\examples{
my.pval(c(0.00000001,0.01,0.9))

}
\seealso{
format.pval
}
\keyword{p-value}

