% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.traceplot.R
\name{my.traceplot}
\alias{my.traceplot}
\title{Custom traceplot with thinning and burn-in}
\usage{
my.traceplot(trace.vec, label = NULL, thin = 0, burn = 0, col = "black")
}
\arguments{
\item{trace.vec}{Vector containing draws from the MCMC sampler}

\item{label}{Parameter label and title for the plot. Default is no label.}

\item{thin}{Amount of thinning applied to the plot.  For example, if thin=3, 
then every third sample point will be plotted.  The total number of 
iterations should be divisible by the thinning amount.}

\item{burn}{Number of iterations for burn-in.}

\item{col}{Color of trace line. Defaults to black.}
}
\description{
Function to plot a single parameter from a single chain of an MCMC.  
The burn-in period will be included by shaded red.  Additional chains can
be added to the same plot using the add.chain() function
}
\examples{
## Draw a traceplot for a random sample
my.samples = rnorm(10000, 5, 10)
my.traceplot(my.samples,label="Some title",thin=5,burn=1000,col="blue")

}
\seealso{
add.chain
}
\keyword{traceplot}

