% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_rows}
\alias{normalize_rows}
\title{Normalize rows to sum to 1}
\usage{
normalize_rows(M)
}
\arguments{
\item{M}{The matrix with rows in need of normalization}
}
\description{
Given a matrix, divide each row by its sum in order to normalize so that
each row sums to 1.  Useful for obtaining category probabilities.
}
\details{
Note: doesn't perform any validation or checking
}

