% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMultinom}
\alias{rMultinom}
\title{Sample from multinomial distribution}
\usage{
rMultinom(probs, m = 1L)
}
\arguments{
\item{probs}{\eqn{N \times K} matrix of probability of observation \eqn{n} falling 
into category \eqn{k}$}

\item{m}{How many outcomes should be sampled per row?  Defaults to 1.}
}
\description{
Given a matrix of category probabilities, return a matrix of outcomes
sampled from multinomial distribution with those category probabilities
}
\details{
Note: doesn't perform any validation or checking
}

