% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{slow_column_picker}
\alias{slow_column_picker}
\title{Choose column based on an input vector}
\usage{
slow_column_picker(M, column)
}
\arguments{
\item{M}{The input matrix}

\item{column}{The vector containing the columns to be extracted}
}
\description{
Choose column based on an input vector
}

