% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dMultinom}
\alias{dMultinom}
\title{Get density of multinomial distribution}
\usage{
dMultinom(probs, y, lg = FALSE)
}
\arguments{
\item{probs}{$N \times K$ matrix of probability of observation $n$ falling
into category $k$}

\item{y}{Vector containing the value at which the density should be evaluated.
Values should range from 1 to $K$}

\item{lg}{Should the log of the density be returned?  Defaults to false}
}
\description{
Given a matrix of category probabilities and a vector of realizations,
return a vector of the multinomial density evaluated at those categories
}
\details{
Note: doesn't perform any validation or checking
}

