% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomFuncts.R
\name{initialize_containers}
\alias{initialize_containers}
\title{Function to initialize containers}
\usage{
initialize_containers(R, n, coef_U, coef_M, coef_Y, thin = 1, imp_U = FALSE)
}
\arguments{
\item{R}{Number of MCMC scans}

\item{coef_U}{Vector containing seed values for U coefficients}

\item{coef_M}{Matrix containing seed values for M coefficients}

\item{coef_Y}{Vector containing seed values for Y coefficients}

\item{thin}{How much to thin container for imputed Us}

\item{imp_U}{Whether to save imputed U's.  Defaults to FALSE.}
}
\description{
Initialize containters for regression coefficients, acceptance, and imputed
}

