% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latexMat.R
\name{latexmat}
\alias{latexmat}
\title{Write LaTeX code to print a matrix}
\usage{
latexmat(mat, matType = "pmatrix")
}
\arguments{
\item{mat}{The matrix to be printed}

\item{matType}{Bracketing type for the matrix. Defaults to "pmatrix"}
}
\description{
Write LaTeX code to print a matrix
}
\examples{
## Make the matrix
a <- rbind(1:4,5:8)

## Apply the print function
latexmat(a)
}

