% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latexMat.R
\name{latexMat}
\alias{latexMat}
\title{Write LaTeX code to print a matrix}
\usage{
latexMat(mat, matType = "pmatrix", fractions = TRUE, digits = 3,
  align = "r")
}
\arguments{
\item{mat}{The matrix to be printed}

\item{matType}{Bracketing type for the matrix. Defaults to "pmatrix"}

\item{fractions}{Should we write the elements in fraction form (uses MASS package)? Defaults to TRUE}

\item{digits}{Number of decimal places to print (if not using fractions)}

\item{align}{Number alignment within columns. Options are "c" for center, "r" for right, "l" for left.
Default is right aligned.}
}
\description{
Write LaTeX code to print a matrix
}
\examples{
## Make the matrix
a <- rbind(1:4, 5:8/10)

## Apply the print function
latexMat(a)
}

