% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.kable.R
\name{my.kable}
\alias{my.kable}
\title{Customized kable (table-making function for knitr)}
\usage{
my.kable(kable.object, booktabs = TRUE, escape = FALSE, format = "latex",
  caption = NULL, ...)
}
\arguments{
\item{booktabs}{Whether booktabs style should be used.  Defaults to TRUE.}

\item{escape}{Whether special LaTeX characters should be escaped.
The default in kable right now is to escape these characters, but for some 
reason, escape never works quite like I want it to. Defaults to FALSE.}

\item{format}{Type of table to be made.  Defaults to "latex".}

\item{caption}{Caption for the table. Defaults to NULL (no caption).}

\item{...}{Other parameters passed to kable}

\item{kable.obj}{Object (matrix or data frame) passed in to be kabled}
}
\description{
Function to adapt kable function from knitr package to meet most of my
table-making needs. This intentionally removes the table environment, so the
tables do not float. Unfortunately, that means it's hit or miss on whether 
the caption stays near the table.  This function is ugly
}
\seealso{
kable
}
\keyword{knitr}
\keyword{tables}

