% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unroll.R
\name{unroll}
\alias{unroll}
\title{Unroll a list and save its elements into the global environment}
\usage{
unroll(list)
}
\arguments{
\item{list}{Name of the list to unroll}
}
\description{
Used for its side effects
}

